EIN-4975: Skip orphan-check if IDs are equal

+++ src/main/java/no/einnsyn/backend/entities/journalpost/JournalpostService.java
@@ -167,7 +167,7 @@ public class JournalpostService extends RegistreringService<Journalpost, Journal
       var oldSkjerming = journalpost.getSkjerming();
       journalpost.setSkjerming(skjermingService.findOrCreate(skjermingField));
 
-      if (oldSkjerming != null) {
+      if (oldSkjerming != null && !oldSkjerming.getId().equals(skjermingField.getId())) {
         skjermingService.deleteIfOrphan(oldSkjerming);
       }
     }
